Remove remaining non-SDL usage of SDL fixed width int types. (#314)
* Remove remaining non-SDL usage of SDL fixed width int types.
* Clean up Wii process_event.

+++ arch/3ds/platform.c
@@ -50,7 +50,7 @@ int pclose(FILE *stream)
   return 0;
 }
 
-void delay(Uint32 ms)
+void delay(uint32_t ms)
 {
   if(ms > 0)
   {
@@ -68,9 +68,9 @@ boolean ctr_supports_wide(void)
   return consoleModelId != 3 /* O2DS */;
 }
 
-Uint32 get_ticks(void)
+uint64_t get_ticks(void)
 {
-  return (Uint32)osGetTime();
+  return osGetTime();
 }
 
 boolean platform_init(void)
